refactor[venom]: introduce `IRContext` and `IRAnalysisCache` (#3983)
this is a refactoring commit. it introduces `IRContext` and
`IRAnalysisCache`. `IRContext` is a global compilation context for a
compilation run, it includes global information like the deploy- and
runtime-code. meanwhile, `IRFunction` now corresponds directly to
"internal" functions passed from the frontend.

this commit refactors the `IRAnalysis` and `IRPass` classes by adding
`IRAnalysisCache` (kind of a manager for analyses) which determines
which analyses are needed for a given pass.

additionally, there is a compilation-time performance improvement which
is due to optimizations running over smaller sections of code.

this commit introduces some performance regressions due to
`remove_unused_variables` running in one-shot instead of in a loop now.
these performance regressions will be investigated and tuned in a
follow-up PR.