fix[venom]: fix `cfg` output format (#4010)
this commit fixes a regression introduced in ace3789e4b9e2307, as we now
have multiple `IRFunction` objects grouped in an `IRContext`. This
commit adds the required `as_graph()` method to `IRContext`. At the same
time the `IRFuntions.as_graph()` is updated to support emitting only a
subgraph without wrapping it into a graphviz `digraph`.