feat[venom]: optimize mem2var and store/variable elimination pass sequences (#4032)
The `Mem2Var` pass needs to operate on SSA form. This commit 
updates the order of passes to accommodate for this. Additionally
we add an extra `StoreElimination` pass after `SCCP` to further
reduce the generated code size.

Additionally, it updates the `RemoveUnusedVariablesPass` to eliminate
all unused variables without the need to be run multiple times.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>