feat[venom]: "pickaxe" stack scheduler optimization (#3951)
implement "pickaxe" stack scheduler optimization.

sometimes, venom produces some unneeded stack traffic. this can't be
fixed by improved dataflow analysis, because it is an issue in the stack
scheduler rather than instruction reordering. what happens is a stack
item is produced, but it has a longer lifespan than other things on the
stack and the instruction cannot be reordered. this commit introduces a
simple heuristic: at the time of stack item production, check the live
list for something with "shorter" liveness we can swap it with.

i just made up the "pickaxe" term but i think it sounds cool :)