feat[lang]!: make internal decorator optional (#4040)
make `@internal` decorator optional. since functions can be imported
now, the `@internal` decorator could be confusing UX, since it is not
clear (without referencing docs) whether `@internal` means "hidden from
importers of this module", or "not included in the selector table". with
this commit, functions don't need a visibility marker, and are just
internal by default.