fix[tool]: compile multiple files (#4053)
fix compilation for multiple files where they initialize the same
module. the analysis was getting cached between compilations, leading to
a compiler panic on allocating the storage layout (because the module
was previously touched by the allocator).

this was not caught in previous testing because the pattern in the test
suite is to run a single compilation per test, with a fresh input
bundle.