fix[codegen]: fix src/dst overlap in dynarray_append
make_setter can potentially run into overlap when called from
dynarray_append; this commit copies to a temporary intermediate buffer
before copying to the destination if the condition is detected.

this commit also adds a util function to detect overlap, and adds an
assertion directly in `make_setter` so that future variants panic
instead of generating bad code.