fix[codegen]: panic on potential eval order issue for some builtins (#4157)
`extract32()` and `slice()` have an evaluation order issue when
the arguments touch the same data. specifically, the length and data
evaluation are interleaved with the index/start/length evaluations. in
unusual situations (such as those in the included test cases), this
can result in "invalid" reads where the data and length reads appear
out of order. this commit conservatively blocks compilation if the
preconditions for the interleaved evaluation are detected.

---------

Co-authored-by: trocher <trooocher@proton.me>
Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>