feat[venom]: add small heuristic for cleaning input stack
this commit adds a small heuristic for cleaning the input stack from
cfg_in, which is to pop the shallowest items first.

it also cleans up the code a little bit and adds stronger preconditions
to the code.

a further optimization would be to try to align the output stack as best
as possible at the jnz, but that belongs in the DFT pass.