feat[tool]: add integrity hash to initcode (#4234)
this commit adds the integrity hash of the source code to the
initcode. it extends the existing cbor metadata payload in the
initcode, so that verifiers can compare the integrity hash to the
artifact produced by a source bundle.

the integrity hash is put in the initcode to preserve bytecode space of
the runtime code.

refactor:
- change existing `insert_compiler_metadata=` flag to the more generic
`compiler_metadata=None`, which is more extensible.