feat[venom]: store expansion pass (#4068)
expand `extract_literals` pass (introduced in bb9129aaa42b2f810f) to
also extract variables and rename it to `store_expansion`, allowing for
the dft pass to reorder variable uses more effectively.

since this also gives us the guarantee that each variable is used
exactly once (besides by store instructions), this allows us to make
some simplifications in `venom_to_assembly.py`, since we no longer need
to account for the same variable occurring on the stack more than one
time (cf. for example 5d8280feec16f86ae).

this results in a stack scheduler improvement. for example:
- examples/tokens/ERC20.vy has a 20 byte codesize improvement
- examples/auctions/blind_auction.vy has a 145 byte codesize improvement
- examples/voting/ballot.vy has a 28 byte codesize improvement

across a range of contracts, the improvement seems to be between 1-2%,
but sometimes as high as 5%

since stack operands are now guaranteed to be unique, the old rule to
avoid swapping if two stack operands are the same no longer works. to
address this, this commit adds an equivalence analysis. this creates
equivalence sets of equivalent variables based on store chains, and
then generalizes the rule from "don't swap if two stack operands are
the same" to "don't swap if two stack operands are equivalent".