feat[lang]: add native hex string literals (#4271)
This commit adds support for custom hex literals like `x"a8b9"`. this
hex literal is equivalent to the bytestring `b"\xa8\xb9"`. this syntax
is a break from Python, which does not have this type of literal, but
it is helpful in the smart contract space because often literals are
copy-pasted around as hex.

the approach taken in this PR is to add another custom sub-parser in
`pre_parse.py` which keeps track of where the hex literals are in
the original source code. it additionally refactors the signature of
`pre_parse()` to return a `PreParseResult`, which bundles all the data
structures returned by `pre_parse()` together.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>