feat[tool]: separate import resolution pass (#4229)
this commit separates import resolution into its own pass. there are
two reasons for this:

a) improved performance on certain operations, specifically,
computation of the integrity hash. this is a hotspot for tooling, which
needs to compute the integrity hash of a contract quickly to check if
it needs to be recompiled or not

b) logical separation. after this commit, all the I/O and filepath
computation of the module analysis is moved into its own pass, making
the module analyzer "pure" - it no longer needs to deal about input
bundles or filepaths or anything like that.

misc/refactor: this commit moves integrity hash computation into
a compiler pass. besides that, some some minor code touchups were
performed, but because of moving code into a new file, an effort
was made to keep the semantics of the code largely the same, to ease
reasoning about the diff.