remove init function from interfaces
it was only there for `-f abi` output -- since it is a standards
requirement to have the constructor in the abi output, but it doesn't
semantically make sense in-language for the init function to be the
interface, we add it back in later, at abi generation time.

add a test that `module.__interface__(...).__init__()` is not allowed.