feat[venom]: reduce legacy opts when venom is enabled (#4336)
reduce legacy IR optimization when venom is enabled. the use of
`IRnode._optimized` was there to decide if it was safe to inline an
IRnode expression or it required a `with` statement; with venom, we
don't need to inline the expressions, since the venom optimizer is more
powerful. this leads to a 50% improvement in AST -> IRnode generation,
which is a ~15% performance improvement in overall end-to-end compile
time.