fix[codegen]: assert staticcall for identity and ecrecover precompiles
in 93a957947af1088addc, the assert for memory copying calls to the
identity precompile was optimized out; the reasoning being that if the
identity precompile fails due to OOG, the caller contract would also
likely fail with OOG. however, due to the 63/64ths rule, there are cases
where the identity precompile could fail with OOG, but the caller
contract has enough gas to successfully return out of the call context.

(note that even prior to 93a957947af1088addc, some calls to the identity
precompile did not check the success flag. cf. commit cf03d27be6a74c0c).

this only affects pre-cancun compilation targets, since post-cancun the
`mcopy` instruction is chosen. this commit fixes the identity precompile
call for pre-cancun targets.

a similar optimization was also applied in the past to omit the check
when calling the ecrecover precompile. a check for the ecrecover call is
applied in this commit.

while this is a performance regression, ecrecover is generally not a
hotspot; as for the identity precompile, it is only used by the compiler
for memory copies pre-cancun, so the performance regression there is not
too relevant (as of nov 2024).