feat[venom]: make cfg scheduler "stack aware"
this is a step towards making the cfg traversal order more "stack
aware". previously we would blindly try to remove `iszero` instructions
before a `jnz`. now, the heuristic is based on which basic block has a
smaller sized dependency on this basic block. only if the two cfg_out
blocks have the same sized dependency, then we try to remove the iszero.