fix[ux]: fix error message for common typo
the `staticall` typo is very common, and was raising a hard-to-read
error message from the python parser like "invalid syntax. Perhaps you
forgot a comma?"

this commit adds it to the pre-parser, so we can raise an exception
after we have successfully gone through the python parser and have the
ability to add source info and a hint.