fix[ux]: fix empty hints in error messages (#4351)
remove empty hint message from exceptions. this regression was
introduced in 03095ce3921636. the root cause of the issue, however,
was that `hint=""` could be constructed in the first place. this commit
fixes the `get_levenshtein_error_suggestions` helper so that it returns
`None` on failure to find a suggestion rather than the empty string
`""`.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>