fix[tool]: add missing user errors to error map  (#4286)
This commit fixes a bug where not all errors are present in the
error_map (in the `-f source_map` output). This could happen when the
`IRnode` with an error message was optimized out, since the error
message was not propagated to its descendants.

To resolve this, two changes were made. In `set_error_msg`, the code
now ensures that existing error messages are not overwritten if they
have already been set. In `IRnode.from_list`, when an `error_msg`
argument is provided, it is assigned to all nodes in the list unless a
node already has an existing `error_msg`. Both changes were made, since
both methods are used in the codebase to set `IRnode.error_msg`.