fix[tool]: add missing internal functions to metadata (#4328)
this commit adds all reachable internal functions to the `-f metadata`
output. this was overlooked during the 0.4.0 release - there could
be internal functions in submodules which didn't get added to the `-f
metadata` output, since the routine to build the `-f metadata` output
only looked at the top-level module.

this commit also makes a change to the format of the `.function_info`
keys - since function names are not guaranteed to be globally unique
anymore (that is, two different modules may both contain a function
named `foo`), this commit adds a disambiguating function id to each
key in `.function_info`. (it uses the function_id computed during IR
generation since it is guaranteed to be unique). the function name
is still present in the function metadata under the `.name` key, but
downstream tooling may have to adapt if they had the assumption that
the keys in `.function_info` were parseable names. for illustration,
here is an example of function metadata prior to and after this commit.

prior to this commit:
```json
{
  "function_info": {
    // valid identifier, but not necessarily unique
    "foo": {
      "name": "foo",
      "_ir_identifier": "internal 0 foo()",
      ...
    }
  }
}
```

after this commit:
```json
{
  "function_info": {
    // not a valid identifier, but is unique. for the function name, see
    // the "name" field below.
    "foo (0)": {
      "name": "foo",
      "_ir_identifier": "internal 0 foo()",
      ...
    }
  }
}
```

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>