refactor[codegen]: add profiling utils
add two new profiling utils, `profileit()` and `cumtimeit()` which run
cProfile and printout cumulative time spent in a context, respectively.

the functions both return a context manager, so it can be used both in a
with statement or as a function decorator, e.g.:
```python
@profileit()
def foo():
    ...

@cumtimeit("foo")
def foo():
    ...
```
in a with statement
```
with profileit():
    foo()

with cumtimeit("foo"):
    foo()
```

neither of these should be used except for in local dev work, so they
are ignored in coverage, and a warning is printed out if they are used
as a reminder to remove them before pushing.