feat[venom]: remove special cases in store elimination (#4413)
remove special cases in store elimination, allowing elimination of
`%ret_ofst` and `%ret_size` store chains. this was previously to
protect correctness in `mem2var`, but it no longer seems to be needed.

running an additional store elimination before `mem2var` results in a
slight bytecode size improvement, since it allows more memory locations
to be promoted to stack variables.