fix[venom]: fix `MakeSSA` with existing phis (#4423)
This commit improves the `MakeSSA` pass to handle incoming `phi`
instructions. Following the implementation of the venom parser it
is possible to parse code with existing `phi` instructions in the
code. The old code was expecting the `phi` instructions to have
been self-placed and in "degenerate form" of output == all branch
arguments. The new code does not make this assumption.

This commit additionally:
- refactors the existing `make_ssa` tests to use new test machinery
- adds a phi parsing test (does not test the new code in this commit
  since it does not run passes, but does make sure we can at least parse
  phis)
- expands the venom round-trip tests to check that we can both
    a) run venom passes on parsed venom, and
    b) bytecode generation from round-tripping venom produced by the
       vyper frontend is equivalent to bytecode generation from the
       regular pipeline (directly from vyper source code)

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>