fix[venom]: fix store elimination pass (#4428)
this commit fixes the store elimination pass by updating the dfg
in-place instead of relying on a stale dfg.

this currently results in no bytecode changes. previously this
was undetected because the order of items in the dfg happens to be
"well-behaved", but if the dfg is built using a traversal of basic
blocks in a different order (as may happen in upcoming passes), it can
result in store instructions failing to be eliminated.

note that we haven't rebuilt the dfg properly because `dfg.outputs`
is invalid after this pass. we could modify `dfg.outputs` in place, but
that results in a bytecode regression.

this commit also removes the dependency on CFGAnalysis as it is not
actually needed by the pass.

---------

Co-authored-by: Harry Kalogirou <harkal@nlogn.eu>