remove VyperNode __hash__ and __eq__ implementations
it is a performance and correctness footgun for the hash and eq
implementations to recurse. for instance, two nodes from different
source files should never compare equal (you can use deepequals to
signal intention there).

this commit temporarily overrides the `__ne__` implementation for
VyperNode so that we can find all the `!=` uses in the tests.