feat[tool]: support storage layouts via `json` and `.vyz` inputs (#4370)
This commit adds support for overriding the storage layout using
`solc_json` and archive inputs, and consequently adding the storage
layout if it was provided to these formats as output. This makes it
possible for verifiers to verify code compiled with a storage layout
override with no changes on their end.

A design decision was made to have the storage layout override affect
the integrity hash. This is so you can tell that a contract was
compiled with storage layout override (even if it does not affect the
bytecode).

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>